feat: Disable external ingress for better security
This commit adds a new check for ACA to assess if external ingress is disabled. The check is implemented in the `isExternalIngressEnabled` function in the `ACACheck` class. The `acaRules.json` file is also updated with the new rule `ACA_Private_DisableExternalIngress` which expects external ingress to be disabled for better security.

Closes #9