feat: Add check for managed identity enabled
This commit adds a new check for ACA to assess if managed identity is enabled. The check is implemented in the `isManagedIdentityEnabled` function in the `ACACheck` class. The `acaRules.json` file is also updated with the new rule `ACA_Identity_UseManagedIdentity` which expects managed identity to be enabled for better security.

Closes #11